Add support for rvalue iterators
Before this commit all iterators returned references (lvalues).
Now it is possible to return either lvalues or rvalues depending
on the template arguments, e.g.:

```C++
  // These will return `float&`:
  SETUP_ITERATORS(myClass, float&, it_state);
  SETUP_REVERSE_ITERATORS(myClass, float&, it_state);

  // These will return `float`:
  SETUP_ITERATORS(myClass, float, it_state);
  SETUP_REVERSE_ITERATORS(myClass, float, it_state);
```

This is reflected on `it_state::get()` functions that are forced
to return the same types as the iterators.

(Note that before this commit if a get function wanted to return
an rvalue it needed to create a temporary variable to store it
and then return a reference to this variable)